v2.1.13
Fixed issues with tvOS and tap gesture recognizer not being available.  Added several @Published and notification checks for things that aren't available on Linux.  Changed alignment of "Total Capacity:" label so it's not floating by itself.  Created constant for `devicePanelRadius` for consistency.  Fixed issue with PreferenceKey monitoring and changed to onAppear and onChange modifiers which is simpler and works better.