v2.8.6
Updated change log with SWIFTPACKAGEINDEX test format to be part of description and not separate lines.
Updated the README to have an updated list of Capabilities to match the new additions in a similar order.
Removed DTS Case-ID: 8753208 since it appears this no longer applies in current version of Xcode.
Moved Apple Pay capability to be first under additional features.
Added initial screenshot feature (likely should be improved like including the status bar) but should work on iOS & macOS & tvOS.